Rewrite custom sso button stuff in light flow to just something for FC
